chore: Move schedules.py to utils/iteration/ (Option C)
Per alg/ audit discussion: schedules are iteration infrastructure but
not convergence observation. utils/iteration/ is a clean separation
with room to grow (callbacks, acceleration wrappers).

- New: utils/iteration/schedules.py
- Removed: utils/convergence/schedules.py
- Updated: 2 import sites + deprecation shim
